:fire: secrets from env vars review part
